Warning fixes
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@35323 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tools/lighter2/lightcalculator.cpp
@@ -158,7 +158,7 @@ namespace lighter
       PrimitiveArray& primArray = submeshArray[submesh];
 
 	  #pragma omp parallel for
-      for (int pidx = 0; pidx < primArray.GetSize (); ++pidx)
+      for (size_t pidx = 0; pidx < primArray.GetSize (); ++pidx)
       {
         // Get next primitive
         Primitive& prim = primArray[pidx];
@@ -200,7 +200,7 @@ namespace lighter
 
         // Iterate all primitive elements
 		#pragma omp parallel for
-        for (int eidx = 0; eidx < numElements; ++eidx)
+        for (size_t eidx = 0; eidx < numElements; ++eidx)
         {
           // Skip empty elements
           Primitive::ElementType elemType = prim.GetElementType (eidx);
